# backend/docs/api/property-management-openapi.yaml
openapi: 3.0.3
info:
  title: PropMetrik Property Management API
  description: |
    Property Management API for PropMetrik platform.
    
    ## Features
    - **Property Management**: Properties, units, amenities, and listings
    - **Tenancy Management**: Lease lifecycle, tenant onboarding, renewals
    - **Payment Processing**: Rent collection via Paystack (cards, mobile money)
    - **Maintenance**: Work order submission and status tracking
    - **Lease Templates**: Template CRUD, variable substitution, PDF generation
    - **Tenant Portal**: Self-service for tenants (payments, maintenance, documents)
    
    ## Authentication
    - Admin/Staff endpoints use Keycloak Bearer tokens
    - Tenant Portal uses magic link/OTP-based authentication
    
    ## Payment Modes
    - Test mode: Uses Paystack test keys (no real transactions)
    - Live mode: Uses Paystack live keys (real transactions)
    
  version: 1.0.0
  contact:
    name: PropMetrik Support
    email: support@propmetrik.com
servers:
  - url: https://api.propmetrik.com/v1/pm
    description: Production server
  - url: https://staging-api.propmetrik.com/v1/pm
    description: Staging server
  - url: http://localhost:3001/api/pm
    description: Local development
...
paths:
...
  /properties:
    get:
      tags:
        - Properties
      summary: List properties
      description: Retrieve paginated list of properties with optional filters
      operationId: listProperties
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, under_maintenance]
        - name: propertyType
          in: query
          schema:
            type: string
            enum: [residential, commercial, mixed_use, land]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Search by property name or address
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []

    post:
      tags:
        - Properties
      summary: Create property
      operationId: createProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePropertyRequest'
      responses:
        '201':
          description: Property created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/BadRequest'
      security:
        - bearerAuth: []
  /properties/{propertyId}:
    get:
      tags:
        - Properties
      summary: Get property details
      operationId: getProperty
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []

    patch:
      tags:
        - Properties
      summary: Update property
      operationId: updateProperty
      parameters:
        - $ref: '#/components/parameters/PropertyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePropertyRequest'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
      security:
        - bearerAuth: []

  # ===== UNITS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token for admin/staff access
    
    tenantToken:
      type: http
      scheme: bearer
      description: Session token for tenant portal access
  parameters:
    PropertyId:
      name: propertyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean

    # Property Schemas
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, mixed_use, land]
        address:
          type: string
        city:
          type: string
        region:
          type: string
        status:
          type: string
          enum: [active, inactive, under_maintenance]
        totalUnits:
          type: integer
        occupiedUnits:
          type: integer
        createdAt:
          type: string
          format: date-time
    PropertyDetail:
      allOf:
        - $ref: '#/components/schemas/Property'
        - type: object
          properties:
            description:
              type: string
            amenities:
              type: array
              items:
                type: string
            images:
              type: array
              items:
                type: string
            owner:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                email:
                  type: string
    PropertyListResponse:
      type: object
      properties:
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreatePropertyRequest:
      type: object
      required:
        - name
        - propertyType
        - address
      properties:
        name:
          type: string
        propertyType:
          type: string
          enum: [residential, commercial, mixed_use, land]
        address:
          type: string
        city:
          type: string
        region:
          type: string
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
    UpdatePropertyRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, inactive, under_maintenance]
        description:
          type: string
        amenities:
          type: array
          items:
            type: string

    # Unit Schemas
...